+++ Runtime/Spells/CorrectSpell.cs
@@ -1,0 +1,11 @@
+using UnityEngine.UI;
+
+namespace UIWitches
+{
+    public class CorrectSpell : UISpell<Dropdown>
+    {
+        public float f;
+        public float[] fs;
+        public int i;
+    }
+}