+++ Runtime/TextMeshPro/Witches/TMP_DropdownWitch.cs
@@ -1,17 +1,9 @@
 using TMPro;
 using UnityEngine;
-using UnityEngine.Events;
 
 namespace UIWitches.TMPro
 {
-    /// <summary>
-    /// A UI Witch for the <c>TMP_Dropdown</c> Component, using <c>ITMP_DropdownSpell</c>.
-    /// </summary>
     [RequireComponent(typeof(TMP_Dropdown))]
     [AddComponentMenu("UI/Witches/Dropdown - TextMeshPro Witch")]
-    public class TMP_DropdownWitch : UIWitch<TMP_Dropdown, int, ITMP_DropdownSpell>
-    {
-        protected override UnityEvent<int> onValueChanged => selectable.onValueChanged;
-        protected override UnityAction<int> setWithoutNotify => selectable.SetValueWithoutNotify;
-    }
+    public class TMP_DropdownWitch : UIWitch<TMP_Dropdown> { }
 }