+++ Runtime/TextMeshPro/Witches/TMP_TextWitch.cs
@@ -1,0 +1,9 @@
+using TMPro;
+using UnityEngine;
+
+namespace UIWitches.TMPro
+{
+    [RequireComponent(typeof(TMP_Text))]
+    [AddComponentMenu("UI/Witches/TextMeshPro - Text Witch")]
+    public class TMP_TextWitch : UIWitch<TMP_Text> { }
+}