+++ Runtime/Witches/TextWitch.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace UIWitches
+{
+    [RequireComponent(typeof(Text))]
+    [AddComponentMenu("UI/Witches/Text Witch")]
+    public class TextWitch : UIWitch<Text> { }
+}